generate files to store vocab and word vectors
This file uses the NLP tool spaCy to generate a vocabulary, and the corresponding word vectors. The word vectors are inserted into a space-partitioning tree, in this case a "balltree", for fast searching. The vocabulary and balltrees are then broken up into a number of small files, and serialized using pickle, a Python tool.